+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.3.1
+* new language entry settings
+    * `combineIdenticalScopes` - if `true`, combine adjacent tokens with identical scopes
+* fix #16
+
 ## 0.3.0
 * Better method to hide substituted text, fixing #4 #5
     * new mode (default): "prettifySymbolsMode.hideTextMethod": "hack-letterSpacing"